Add utils to PermissionTree
